game: warning fix

+++ src/game/bg_animation.c
@@ -1648,7 +1648,7 @@ int BG_GetConditionValue(int client, int condition, qboolean checkConversion)
 		else
 		{
 			// must use COM_BitCheck on the result.
-			return (int)globalScriptData->clientConditions[client][condition];
+			return (intptr_t)globalScriptData->clientConditions[client][condition];
 		}
 		//BG_AnimParseError( "BG_GetConditionValue: internal error" );
 	}
